refactor: ViewResolver 인터페이스 추가 및 정적 리소스 view resolver 구현체 추가
