+++ tomcat/src/main/java/org/apache/coyote/view/ViewResolver.java
@@ -1,0 +1,8 @@
+package org.apache.coyote.view;
+
+import org.apache.coyote.response.HttpResponse;
+
+public interface ViewResolver {
+
+    void resolve(String fileName, HttpResponse response);
+}
